project structuring + setup auth routes and controller
